move base / root types around
