+++ rpc/types/README.md
@@ -57,6 +57,6 @@ values inside JSON strings, for example:
 }
 ```
 
-`binary` here is (de)serialized as a normal [`String`]. In order to be clear on which fields contain binary data, the struct fields that have them will use [`crate::data::BinaryString`] instead of [`String`].
+`binary` here is (de)serialized as a normal [`String`]. In order to be clear on which fields contain binary data, the struct fields that have them will use [`crate::BinaryString`] instead of [`String`].
 
 TODO: list the specific types.
