--- rpc/types/src/base/empty_request_base.rs
@@ -1,23 +1,0 @@
-//! TODO
-
-//---------------------------------------------------------------------------------------------------- Import
-use serde::{Deserialize, Serialize};
-
-//---------------------------------------------------------------------------------------------------- EmptyRequestBase
-/// TODO
-///
-/// <https://github.com/monero-project/monero/blob/cc73fe71162d564ffda8e549b79a350bca53c454/src/rpc/core_rpc_server_commands_defs.h#L95-L99>.
-#[derive(
-    Copy, Clone, Debug, Default, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize,
-)]
-pub struct EmptyRequestBase;
-
-cuprate_epee_encoding::epee_object! {
-    EmptyRequestBase,
-}
-
-//---------------------------------------------------------------------------------------------------- Tests
-#[cfg(test)]
-mod test {
-    // use super::*;
-}
