--- rpc/types/src/data/mod.rs
@@ -1,7 +1,0 @@
-//! Data structures that appear in other types.
-//!
-//! TODO
-
-mod binary_string;
-
-pub use binary_string::BinaryString;
